requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
Django==4.0
django-cors-headers==4.3.1
django-el-pagination==4.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.6
lxml==5.1.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.2.0
pillow==10.2.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
